Changment of cursor color value to keep the same color than before VTK8
With VTK update, the color of the cursor was white
I found that, passing value from int (base 256) to double (base 1.0) change the behavior
So I adapted the color to keep the red for the activated image, the green, unactivated slice and light blue for unactivated linked image

+++ vv/vvSlicer.h
@@ -164,7 +164,7 @@ public:
   void SetCursorVisibility(bool s);
 
   bool GetCursorVisibility();
-  void SetCursorColor(int r,int g, int b);
+  void SetCursorColor(double r,double g, double b);
 
   void SetCornerAnnotationVisibility(bool s);
   bool GetCornerAnnotationVisibility();
